requirements.txt
Django==1.7
django-twitter-api
django-suit